+++ scripts/split-e2e-tests.ts
@@ -132,17 +132,18 @@ const RUN_SOLO: (string | RegExp)[] = [
   /src\/__tests__\/deploy-velocity\/.*\.test\.ts/,
   // SQL tests
   /src\/__tests__\/rds-.*\.test\.ts/,
+  /src\/__tests__\/sql-.*\.test\.ts/,
   // CDK tests
   /src\/__tests__\/base-cdk.*\.test\.ts/,
   'src/__tests__/amplify-table-1.test.ts',
   'src/__tests__/amplify-table-3.test.ts',
+  'src/__tests__/amplify-table-4.test.ts',
   'src/__tests__/api_canary.test.ts',
-  'src/__tests__/sql-models-1.test.ts',
-  'src/__tests__/sql-models-2.test.ts',
-  'src/__tests__/sql-pg-models.test.ts',
   'src/__tests__/amplify-table-2.test.ts',
   'src/__tests__/admin-role.test.ts',
   'src/__tests__/all-auth-modes.test.ts',
+  'src/__tests__/default-ddb-canary.test.ts',
+  'src/__tests__/amplify-ddb-canary.test.ts',
 ];
 
 const RUN_IN_ALL_REGIONS = [
@@ -227,7 +228,11 @@ const splitTests = (baseJobLinux: any, testDirectory: string, pickTests?: (testS
 
       if (RUN_SOLO.find((solo) => test === solo || test.match(solo))) {
         if (RUN_IN_ALL_REGIONS.find((allRegions) => test === allRegions || test.match(allRegions))) {
-          testRegions.forEach((region) => {
+          const shouldRunInNonOptInRegion = RUN_IN_NON_OPT_IN_REGIONS.find(
+            (nonOptInTest) => test.toLowerCase() === nonOptInTest || test.toLowerCase().match(nonOptInTest),
+          );
+          const regionsToRunTest = shouldRunInNonOptInRegion ? nonOptInRegions : testRegions;
+          regionsToRunTest.forEach((region) => {
             const newSoloJob = createJob(os, jobIdx, true);
             jobIdx++;
             newSoloJob.tests.push(test);