+++ packages/graphql-auth-transformer/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [7.2.78](https://github.com/aws-amplify/amplify-category-api/compare/graphql-auth-transformer@7.2.77...graphql-auth-transformer@7.2.78) (2024-04-11)
+
+**Note:** Version bump only for package graphql-auth-transformer
+
 ## [7.2.77](https://github.com/aws-amplify/amplify-category-api/compare/graphql-auth-transformer@7.2.76...graphql-auth-transformer@7.2.77) (2024-03-28)
 
 **Note:** Version bump only for package graphql-auth-transformer