+++ packages/graphql-dynamodb-transformer/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [7.2.76](https://github.com/aws-amplify/amplify-category-api/compare/graphql-dynamodb-transformer@7.2.75...graphql-dynamodb-transformer@7.2.76) (2024-04-11)
+
+**Note:** Version bump only for package graphql-dynamodb-transformer
+
 ## [7.2.75](https://github.com/aws-amplify/amplify-category-api/compare/graphql-dynamodb-transformer@7.2.74...graphql-dynamodb-transformer@7.2.75) (2024-03-28)
 
 **Note:** Version bump only for package graphql-dynamodb-transformer